+++ tests/Mock/Coffee.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace BSP\Tests\Mock;
+namespace BSP\CommandBus\Tests\Mock;
 
 final class Coffee
 {