+++ tests/Mock/ServeCoffeeInCup.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace BSP\Tests\Mock;
+namespace BSP\CommandBus\Tests\Mock;
 
-use BSP\Command;
+use BSP\CommandBus\Command;
 
 final class ServeCoffeeInCup implements Command
 {